Chat_project/kafka_consumer/requirements.txt
kafka-python==2.0.2
streamlit==1.29.0
streamlit_server_state==0.17.1
